Consolidate localtracking into the tractography module.
