+++ AFQ/tests/test_tractography.py
@@ -11,7 +11,7 @@ from AFQ.utils.testing import make_tracking_data
 
 
 seeds = np.array([[ -80., -120.,  -60.],
-                  [-81, -120, -61]])
+                  [-81, -121, -61]])
 
 tmpdir = nbtmp.InTemporaryDirectory()
 fbval = op.join(tmpdir.name, 'dti.bval')
